requirements.txt
iniconfig==2.1.0
packaging==25.0
pillow==11.2.1
pluggy==1.5.0
pygame==2.6.1
pytest==8.3.5
